Update jest tests

+++ jest/teardown.js
@@ -1,7 +1,9 @@
-const selenium = require('selenium-standalone');
-
+const DEV_ENV = process.env.NODE_ENV && process.env.NODE_ENV.toLowerCase().includes('dev');
 
 module.exports = async function () {
-    console.log('Killing selenium server ...'.bgBlue)
-    await global.SELENIUM_SERVER.kill();
+    if (DEV_ENV) {
+        console.log('Killing selenium server ...'.bgBlue);
+        await global.SELENIUM_SERVER.kill();
+        console.log('Killed selenium server'.bgGreen);
+    }
 }
